--- a/mmcv/ops/csrc/common/cuda/correlation_cuda.cuh
+++ b/mmcv/ops/csrc/common/cuda/correlation_cuda.cuh
@@ -29,8 +29,9 @@ using namespace torch;
 #define TensorAcc5R PackedTensorAccessor32<scalar_t, 5, RestrictPtrTraits>
 #define WITHIN_BOUNDS(x, y, H, W) (x >= 0 && x < H && y >= 0 && y < W)
 
-#define THREADS_FORWARD 32
+#define WARP_SIZE 32
 #define THREADS_BACKWARD 16
+#define FULL_MASK 0xffffffff
 
 template <typename scalar_t>
 __global__ void correlation_forward_cuda_kernel(
@@ -42,8 +43,8 @@ __global__ void correlation_forward_cuda_kernel(
   const int C = rInput1.size(3);
 
   const int n = blockIdx.x;
-  const int h = blockIdx.y;
-  const int w = blockIdx.z;
+  const int h = blockIdx.y * blockDim.y + threadIdx.y;
+  const int w = blockIdx.z * blockDim.z + threadIdx.z;
   const int thread = threadIdx.x;
 
   const int start_i = -padH + h * dH;
@@ -52,13 +53,11 @@ __global__ void correlation_forward_cuda_kernel(
   const int patchRadH = dilation_patchH * (patchH - 1) / 2;
   const int patchRadW = dilation_patchW * (patchW - 1) / 2;
 
-  __shared__ scalar_t prod_sum[THREADS_FORWARD];
-
   for (int ph = 0; ph < patchH; ++ph) {
     int ph_dilated = ph * dilation_patchH - patchRadH;
     for (int pw = 0; pw < patchW; ++pw) {
       int pw_dilated = pw * dilation_patchW - patchRadW;
-      prod_sum[thread] = 0;
+      scalar_t prod_sum = 0.0f;
       for (int i = 0; i < kH; ++i) {
         int i1 = start_i + i * dilationH;
         int i2 = i1 + ph_dilated;
@@ -69,23 +68,20 @@ __global__ void correlation_forward_cuda_kernel(
               int j2 = j1 + pw_dilated;
               if
                 WITHIN_BOUNDS(j1, j2, iW, iW) {
-                  for (int c = thread; c < C; c += THREADS_FORWARD) {
+                  for (int c = thread; c < C; c += WARP_SIZE) {
                     scalar_t v1 = rInput1[n][i1][j1][c];
                     scalar_t v2 = rInput2[n][i2][j2][c];
-                    prod_sum[thread] += v1 * v2;
+                    prod_sum += v1 * v2;
                   }
                 }
             }
           }
       }
       // accumulate
-      __syncthreads();
+      for (int offset = 16; offset > 0; offset /= 2)
+        prod_sum += __shfl_down_sync(FULL_MASK, float(prod_sum), offset);
       if (thread == 0) {
-        scalar_t reduce_sum = 0;
-        for (int index = 0; index < THREADS_FORWARD; ++index) {
-          reduce_sum += prod_sum[index];
-        }
-        output[n][ph][pw][h][w] = reduce_sum;
+        output[n][ph][pw][h][w] = prod_sum;
       }
     }
   }
@@ -119,10 +115,9 @@ __global__ void correlation_backward_cuda_kernel_input1(
   const int min_h = h_2 - kH * dilationH;
   const int min_w = w_2 - kW * dilationW;
 
-  __shared__ scalar_t prod_sum[THREADS_BACKWARD][THREADS_BACKWARD];
-  prod_sum[ph_off][pw_off] = 0;
+  scalar_t prod_sum_val = 0.0f;
 
-  for (int ph = ph_off; ph < patchH; ph += THREADS_BACKWARD) {
+  for (int ph = ph_off; ph < patchH; ph += WARP_SIZE) {
     int i1 = h + dilation_patchH * (ph - patchRadH);
     for (int pw = pw_off; pw < patchW; pw += THREADS_BACKWARD) {
       int j1 = w + dilation_patchW * (pw - patchRadW);
@@ -136,8 +131,7 @@ __global__ void correlation_backward_cuda_kernel_input1(
             if (j2 * dW != w_3) continue;
             if
               WITHIN_BOUNDS(i2, j2, H, W) {
-                prod_sum[ph_off][pw_off] +=
-                    grad_output[n][ph][pw][i2][j2] * val;
+                prod_sum_val += grad_output[n][ph][pw][i2][j2] * val;
               }
           }
         }
@@ -145,14 +139,20 @@ __global__ void correlation_backward_cuda_kernel_input1(
     }
   }
 
+  // warp reduce
+  __shared__ scalar_t prod_sum[THREADS_BACKWARD];
+  for (int offset = 16; offset > 0; offset /= 2)
+    prod_sum_val += __shfl_down_sync(FULL_MASK, float(prod_sum_val), offset);
+  if (ph_off == 0) {
+    prod_sum[pw_off] = prod_sum_val;
+  }
+
   __syncthreads();
 
   if (ph_off == 0 && pw_off == 0) {
     scalar_t reduce_sum = 0;
-    for (int ph = 0; ph < THREADS_BACKWARD; ++ph) {
-      for (int pw = 0; pw < THREADS_BACKWARD; ++pw) {
-        reduce_sum += prod_sum[ph][pw];
-      }
+    for (int pw = 0; pw < THREADS_BACKWARD; ++pw) {
+      reduce_sum += prod_sum[pw];
     }
     grad_input1[n][c][h][w] = reduce_sum;
   }
@@ -183,10 +183,9 @@ __global__ void correlation_backward_cuda_kernel_input2(
   const int ph_off = threadIdx.x;
   const int pw_off = threadIdx.y;
 
-  __shared__ scalar_t prod_sum[THREADS_BACKWARD][THREADS_BACKWARD];
-  prod_sum[ph_off][pw_off] = 0;
+  scalar_t prod_sum_val = 0.0f;
 
-  for (int ph = ph_off; ph < patchH; ph += THREADS_BACKWARD) {
+  for (int ph = ph_off; ph < patchH; ph += WARP_SIZE) {
     int i1 = h - dilation_patchH * (ph - patchRadH);
     for (int pw = pw_off; pw < patchW; pw += THREADS_BACKWARD) {
       int j1 = w - dilation_patchW * (pw - patchRadW);
@@ -207,8 +206,7 @@ __global__ void correlation_backward_cuda_kernel_input2(
               if (j2 * dW != w_3) continue;
               if
                 WITHIN_BOUNDS(i2, j2, H, W) {
-                  prod_sum[ph_off][pw_off] +=
-                      grad_output[n][ph][pw][i2][j2] * val;
+                  prod_sum_val += grad_output[n][ph][pw][i2][j2] * val;
                 }
             }
           }
@@ -216,14 +214,20 @@ __global__ void correlation_backward_cuda_kernel_input2(
     }
   }
 
+  // warp reduce
+  __shared__ scalar_t prod_sum[THREADS_BACKWARD];
+  for (int offset = 16; offset > 0; offset /= 2)
+    prod_sum_val += __shfl_down_sync(FULL_MASK, float(prod_sum_val), offset);
+  if (ph_off == 0) {
+    prod_sum[pw_off] = prod_sum_val;
+  }
+
   __syncthreads();
 
   if (ph_off == 0 && pw_off == 0) {
     scalar_t reduce_sum = 0;
-    for (int ph = 0; ph < THREADS_BACKWARD; ++ph) {
-      for (int pw = 0; pw < THREADS_BACKWARD; ++pw) {
-        reduce_sum += prod_sum[ph][pw];
-      }
+    for (int pw = 0; pw < THREADS_BACKWARD; ++pw) {
+      reduce_sum += prod_sum[pw];
     }
     grad_input2[n][c][h][w] = reduce_sum;
   }