@MeowZheng 2022-03-30
env info:
device: v100
gpu driver: 460.73.01
cuda 10.2
pytorch: 1.8.0

eta at 10k iter (total 1200k)
| mmcv master | #1814  | 
| ----  | ----  | 
| 2 days, 21:24:44 |  2 days, 20:44:27 |